TPA-708: merge tpaexec-ee repo to tpa-internal
Change to the github workflow that builds tpa-ee docker image to use the
new folders (tpa-ee/aap24/{rhel,alpine})instead of tpaexec-ee

Changes to the execution-environment.yml files to be used from their new
location

Changes to dockerignore file to reduce weight of docker image

Changes to the script.sh that was not being actively used, it now can be
used to generate an image for both base-image.

Add README.md files to give user more information about the usage of
these files.