Support for flexible M1 architecture
The M1 architecture now supports the following arguments to `tpaexec
configure`:

--primary-location
--data-nodes-per-location
--witness-only-location
--single-node-location

and interprets the "--locations" argument suitably to give any number of
active locations.

A single barman server is set up in the primary location. This node is
also a witness node if either there are no witness-only or single-node
locations, and either there are exactly two locations or there is
exactly one location with an even number of active nodes.

A node given the "witness" role in config.yml will be deployed as a
repmgr witness or EFM witness, according to the cluster's failover
manager.