Add integration tests for flexible M1
New workflows "M1 integration test" and "M1 layout tests".

"M1 integration test" is a variant of the single integration test workflow
but with extra inputs to cover the flexible M1 options.  "M1 layout tests"
runs the integration tests for the combinations of the 6 standard layouts
and {repmgr, efm}, plus patroni with a single location.

To support these, the action that initialises the test now accepts more
inputs and calls `tpaexec configure` appropriately.