Have `configure` define a user-defined docker network and fixed IPs
Docker now uses the same logic to define a random subnet in the
specified network as AWS, but uses one subnet for all locations rather
than one per location as is the 'default'. Changes made:

Add a `hosts_in_cidr` static method to Architecture that returns an
iterator that yields hosts in the given CIDR range

Add a `num_subnets` method to architecture that implements
platform-specfic logic as to how many subnets are needed.

Modify docker.py to declare a Docker network using the name of the
cluster and the randomly generated subnet and add a `networks` entry to
each instance