Add workflow for linting release notes files in PRs
This commit adds the `validate_relnotes.yml` workflow. Its main purpose is to
report missing release notes when PRs are created. This is the complete set
of outcomes that you can get from this workflow:

1. If release notes are missing: send a warning comment in the PR;
2. If release notes are present and valid: send an info comment in the PR;
3. If release notes are present but invalid: finish execution with an
   exception.

We only fail when release notes are present but invalid because that is the
only scenario where we know for a fact that there is an issue.

When release notes are missing, that is not necessarily an issue. Some PRs
do not perform customer-facing changes for example, which would not require
release notes. With that in mind, we avoid failing the job in that case, so
we don't create an unneed blocker on the PR.

For this workflow to work, the repository maintainer needs to create a GitHub
Actions variable for the repository:

- `RELNOTES_DIRECTORY`: it should contain the name of the release notes
  directory inside the source code repository.

Note: the workflow is run only when opening or re-opening a PR. We did it that
way to avoid flooding the PR history with comments when pushing commits to an
existing PR.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>