Install `edb-patroni` instead of `patroni` package
We recently changed the EDB packaging of Patroni. Instead of rebuild the specs
used by PGDG, we are now building our own `edb-patroni` package. This package
has a few differences when compared to PGDG packages:

* The name is `edb-patroni` instead of `patroni` + `patroni-etcd` + ...;
* Uses EDB Python instead of system's Python;
* Uses a virtual environment;
* Brings all the Python dependencies along with Patroni in the package.

Although they have these differences, they offer the same applications under
`/usr/bin` and the same systemd unit under `/lib/systemd/system`.

In that sense, the only differences that TPA should need to handle are:

* The package name: install `edb-patroni` instead of `patroni` package;
* The package source repo: install from EDB repos instead of from PGDG, as the
  package is only available through our repos instead of through both ours and
  PGDG.

This commit applies the above couple of changes to TPA, so it installs the
correct package from the correct repository.

References: TPA-725.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>