Install PGDG repos on etcd nodes when deploying Patroni
Currently the only supported DCS for Patroni is etcd.

When deploying on Debian/Ubuntu we use the outdated `etcd` package from the
system repos.

When deploying on RedHat we use the up-to-date `etcd` package which is
available through PGDG extras.

With that in mind, through this commit we change the `tpaexec configure`
command to automatically add `PGDG` to `yum_repositories_list` of etcd
nodes when deploying a Patroni cluster.

Note: currently TPA does not install PGDG extras on SLES. In any case, we
already changed the logic to also consider `suse_repositories_list`, so once
TPA starts enabling PGDG extras on SLES, we will be ready from Patroni/etcd
perspective.

Note: we only have before-hand information about the operating system when
deploying to Docker or AWS, not when deploying to Bare, which is detected at
deploy time. With that in mind, we are simply setting both `yum_repository_list`
and `suse_repository_list` independently of the underlying operating system. In
the end, that has no effect on other operating systems.

References: TPA-725.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>