Task selector fixes
The last remaining deploy-time tags are now removed in favour of task
selectors, as are a few holdouts in the test roles. In particular, it's
now possible to run only the post-deploy hook with the 'post-deploy'
task selector.

Also some fixes to the task selector documentation to make the examples
work.

Fixes: TPA-713