Gracefully handle "patroni" package replacement with "edb-patroni" package
This commit is introduced to handle existing Patroni clusters, which were
deployed using the "patroni" package (either from PGDG or from EDB
repositories).

As we are now pushing for our "edb-patroni" package, there are a few steps
that need to be performed as part of the package replacement:

1. Detect "edb-patroni" will be installed over "patroni";
2. Pause the Patroni cluster, so we avoid failovers;
3. Stop the Patroni service, as it is using the binary and modules from the
   "patroni" package;
4. Install "edb-patroni" -- this will uninstall "patroni" as well;
5. Start Patroni service, which will now use the binary and modules from the
   "edb-patroni" package;
6. Unpause the Patroni cluster, so we resume to normal operations.

References: TPA-725.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>