Prevent removal of non x86_64 hashes
The hashes of all python modules we compile and ship to customers are
kept in a private repository, not in public pypi. In the same time,
"dependabot" is checking the versions upgrades for us, hence there's no
need to upgrade again (that's why `-U` was removed) but also, and
probably more important,  if we ask `-pip-compile` to upgrade to the
latest versions, it will remove those new hashes because those are
not in the public side.

At the moment, "dependabot" hasn't deleted any private hash.

`--reuse-hashes` is added to improve speed whilst generating its output.