Support both patroni and edb-patroni packages This commit removes the enforcement on using `edb-patroni` packages.
From now on, users are able to define if they want `patroni` (from PGDG) or
`edb-patroni` (from EDB repositories) to be installed. That's defined through
the new `patroni_package_flavour` setting. A corresponding flag was added to
the CLI (`--patroni-package-flavour`). The values made available are `edb`
and `community`, and the default value is `edb`.

The main purpose of this ticket, left aside the flexibility of choosing between
`patroni` and `edb-patroni` packages, is to avoid breaking the deployment of
Patroni clusters by the community, as users who are not subscribed to EDB repos
would not be able to install `edb-patroni`.

It is worth it noticing that migrations between `edb-patroni` and `patroni`
packages are handled gracefully and transparently. At any moment the user can
change `patroni_package_flavour` in the `config.yml` and run
`provision`/`deploy` again. TPA will check if the currently installed package
matches the expectations of the configured value, and if not it will
transparently migrate from one to another.

References: TPA-725.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>