Add support for `postgres_wal_dir` in Patroni deployments
Previous to this commit, if a cluster was deployed by TPA with a custom
`postgres_wal_dir` value, that would not be set in the Patroni YAML file.

That's usually not a problem as TPA takes care of bootstrapping both the
primary and the standby nodes. However, if a standby ever crashes, Patroni
would rebuild the node with `pg_wal` under PGDATA instead of somewhere else
with a symlink.

This commit fixes that issue by adding `postgresql.basebackup.waldir` parameter
to the Patroni YAML file. If Patroni ever needs to rebuild a standby node, it
will make sure to pass `--waldir` parameter to `pg_basebackup` command.

We add `--waldir` both to `bootstrap.postgresql.initdb` and
`postgresql.basebackup` just for completeness. The former will never be used by
Patroni as TPA takes care of the initial bootstrapping of the cluster.

References: TPA-741.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>