Support for the beacon agent
The new role beacon-agent, which can be applied to any postgres node,
deploys the beacon agent and configures it to connect to the beacon
server and to run as a service. A new OS user and database user, both
called "beacon", are created on the node.

The role is applied to postgres servers if `--enable-beacon-agent` is
given to `tpaexec configure`.

The agent requires a beacon project id, read from 'beacon_project_id'
in cluster_vars; this can be passed to `tpaexec configure` with the
`--beacon-agent-project-id` argument. It also requires a beacon access key,
which can be passed in TPA's environment as BEACON_AGENT_ACCESS_KEY
at provision-time, or later by running `tpaexec store-password <clustername>
beacon_agent_access_key`.