Configure `on_role_change` callback in Patroni to reconfigure PgBouncer
This commit adds a `on_role_change.sh` script. This script is installed in the
Patroni nodes and configured as a `on_role_change` callback for TPA clusters
which also contain PgBouncer nodes.

This script is basically a copy and paste of the script used by the integration
between PgBouncer and repmgr, and the idea is the very same one: whenever there
is a failover in the Postgres cluster, update the PgBouncer configuration in
every PgBouncer node to pool connections for the new primary.

Note that the `on_role_change` callback in Patroni is executed both in response
to failovers and switchovers, so both cases are covered by this integration.
The only issue with failovers is: if the PgBouncer nodes are unreachable, their
configuration won't be updated.

References: TPA-754.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>