Add `--enable-pgbouncer` command-line option
This commit introduces the `--enable-pgbouncer` command-line option. This
option was made mutually exclusive with `--enable-haproxy`, the latter
being exclusive to Patroni at this moment.

The idea is to allow the user to select a connection "frontend". Either
HAProxy or PgBouncer can be used (or none of them), depending on which
option the user chooses when running the `tpaexec configure` command.

When `--enable-pgbouncer` is passed to the `configure` command, TPA will
add the `pgbouncer` role to the Postgres nodes, and will configure PgBouncer
to pool connections for the primary node only.

References: TPA-754.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>