Wait in first backup when backing up a replica
When taking a first barman backup from a replica, we force a switch to
a new WAL segment. If we do this too quickly after starting the backup,
then barman will still be backing up the data files and will need the
new WAL segment as well as the old one. So after asking for the backup,
we check on its status and wait for it to reach WAITING_FOR_WALS before
we force the switch.

Because a backup from a replica might still have to wait for the next
barman-cron run to have the WAL moved into the correct directory, the
default backup_wait_timeout is increased to 90 seconds on a replica.

Fixes: TPA-719