Changes related to efm properties
Added text to explicitly tell the user that TPA is setting the
auto.allow.hosts and stable.nodes.file properties to true so that agents
can join the cluster without a separate step needed.

Also removed some example properties that might not have been good
options (see below). To have some example properties, I've added the
property that limits notifications to only warnings and above to cut
down on noise, and ping.serverp that a user should ideally (but not
necessarily) set to some WKA in their environment.

In case anyone is curious, this is why I've removed these properties:

- standby.restart.delay: This would add a delay to restarting every
standby after a promotion, which just makes failovers take longer. I'm
not convinced there's any *good* use case for this property, but the use
case that prompted it was a customer with twenty standbys in a cluster
and they didn't want them all restarted at the same time after a
promotion.
- application.name: This property only makes sense when standbys use an
application name in primary_conninfo, which isn't set up by default by
TPA. If standbys are using it, this property only makes sense when the
nodes use different names. By setting it here, any node that used to be
a primary will have application_name set during a switchover, leading to
a confusing mix of some standbys with the specified name and some
without.
- reconfigure.sync.primary: It's generally bad to take the cluster out
of synchronous mode without a good reason, so it's probably best not
to suggest it.
- reconfigure.num.sync and reconfigure.num.sync.max: It's fine to use
these, but they only make sense when there is more than one standby (and
a max value of less than two doesn't make sense because the lowest it
can be is one).