Run the upgrade-conf command as efm user
On new clusters only, we now generate the config files (.properties and
.nodes) for efm in a subfolder (<efm_config_path>/raw/) before using the
upgrade-config command to decorate the file with comments and inline
documentation in the chosen <efm_config_path> folder.

On existing cluster, we detect that the .properties file exists but not
the /raw folder to know that we shouldn't try to decorate the file.
Since trying to decorate the file in place means every deploy will
regenerate the config from template and then use upgrade-config command
triggering a restart of efm.