Correctly handle failover_manager override for SO nodes and replicas
099f65d34e8c5d8a622e2ea4fd2055a455b7ed57 added support for accepting
repmgr as a valid option for failover_manager for PGD clusters with the
limited use that allows installation of repmgr and subsequent
configuration when HA setup is required for a subscriber-only node. This
however didn't work as intended for clusters using PGD version 5+. This
change intends to fix just that.

References: TPA-767
Signed-off-by: Muhammad Haroon <muhammad.haroon@enterprisedb.com>