Use `create-or-update-comment` in `validate_relnotes.yml`
Now that `create-or-update-comment` exists in `repo-actions`, we can use that
action to maintain a “pinned” comment in each PR that is opened against the TPA
repository.

So far we have been running the `validate_relnotes.yml` workflow only on
`opened`/`reopened` PR events to avoid flooding the PRs with comments as the
current implementation appends a comment each time it is run.

Besides that, we were avoiding to mark the run as "failed" in Actions because
the workflow would be triggered only on `opened`/`reopened` events, and that
would prevent the PR from being merged because of failed checks.

With this commit we intend to use the `create-or-update-comment` action to
create a comment when the PR is opened, and keep updating the comment when the
PR is synchronized or reopened.

That behavior also allows us to mark the workflow run as a failed check if we
hit either of these situations:

* Invalid release notes files are found in the folder;
* Unexpected files are found in the folder.

We still avoid failing when the PR introduces no release notes files because
some PRs are non-customer facing changes, so wouldn't require a release note.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>