Add efm-bind-by-hostname option for M1 + efm
Add the option to use hostnames instead of ip addresses when configuring
efm cluster. `--efm-bind-by-hostname` defaults to false when omitted.

Ensure that the option is only available for M1 + efm clusters.

Set the `bind.address` entry to use the hostname of the node instead of
ip address. Set the .nodes file for efm to use hostnames.

Add doc entry for the option.

Add relnote entry.

Black formatting

References: TPA-758