Preliminary work to allow using existing Barman node as backup node for new clusters
The change allows a cluster to point to an existing Barman node by
entertaining `barman_shared` boolean setting (false by default). A
cluster that wants to reuse an existing Barman node will mark it shared
and the config should look something like this:

    - Name: myBarman
      node: 5
      role:
        - barman
      platform: bare
      ip_address: x.x.x.x
      vars:
          barman_shared: true

Since it is an existing instance, it should be set to use `bare` except
maybe in the first cluster that provisioned and deloyed the instance
itself.

Since some of the configuration would already exist on the Barman node
as part of the first deployment, the change allows TPA to skip trying to
create the same configuration again e.g. the OS user and the
corresponding directories as it may lead to a usermod error with
processes running from the earlier installation.

Also, we don't want to re-create SSH keys for Barman in shared Barman
server. The SSH key pair was created by the first Barman deployment. For
the second and subsequent deployments sharing the Barman node, beside
setting `barman_shared: true`, users also need to copy the
`<first_cluster_dir>/keys/id_barman*` files to the
`<second_cluster_dir>/keys/` directory to allow for the newer clusters
to have access to Barman OS user without breaking the access for the
existing clusters.

Update `~barman/.pgpass` file writting routine so the entries are
prefixed by the pghost rather than the current format which uses '*'.
This does not work for newer clusters reusing an exisitng Barman node
since `barman` and `streaming_barman` Postgres users in the subsequent
clusters are created with different passwords. However, the pgpass
routine always adds rows like these:

```
*:*:*:barman:<password1>
*:*:*:streaming_barman:<password2>
```

And for a shared Barman server we need something like this:

```
pghost1:*:*:barman:<password1>
pghost1:*:*:streaming_barman:<password2>
pghost2:*:*:barman:<password3>
pghost2:*:*:streaming_barman:<password4>
```

This commit introduces an optional pgpass_host parameter to the
postgres/pgpass role. This allows for co-existence of barman and
streaming_barman passwords from multiple Postgres upstream instances
within a single ~barman/.pgpass file. By default, the pgpass_host is set
to '*' to maintain backward compatibility.

There are still a number of areas that require caution when trying to
use a shared Barman instance. Details on the issues are documented in
the barman.md page.

Refere to barman.md for further details on the change and usage
instructions under section 'Shared Barman server'.

Includes minor reformatting at a few different places.

References: TPA-777

Co-authored-by: Muhammad Haroon <muhammad.haroon@enterprisedb.com>
Signed-off-by: Muhammad Haroon <muhammad.haroon@enterprisedb.com>
Signed-off-by: William Ivanski <william.ivanski@enterprisedb.com>