Automate PPC wheels when needed
Bump render template from 1.4 to 1.10
Bump action create PR from 6 to 7

Add -U back

Add a new script that will check the specific hashes for ARM and PPC we
currently use against the newly generated requirement files during
update requirements workflow.

Use either github.head_ref (when run from PR) or github.ref if not available
(manual run) in templating message for the body of the PR.

Ensure that the hash check step fails when we found at least one issue.
If anything is printed by the script and stored inside the ERRORS var
then we know that the check failed.
It will allow the PR to be opened in Draft mode and body of the PR will
be used to list the unmatched dependencies.

remove arm64 file from default targets

add comments

Change upload to only bring requirements.txt instead of full tpa_code

change to include .in files in PR

Modify the parse_hashes script to add the missing hashes when
a match is found for the version and the dep name.

add a template to render the modified requirements file with missing
hashes

Fix the .in file to have an empty line at the end
the file needs to have an empty line at the end of it to be correctly
handled by the script and not ignore the last line.

Update default body for commit and PR desc

remove debug output
fix comments to match latest code
revert to only run in case of merged PR or workflow dispatch event.