Ignore proxy settings when accessing the Patroni API
The Ansible default is to use a proxy, if defined. This does not work in
the (rather common) case of an airgapped environment that needs a proxy
to download packages from the internet, because the proxy also
intercepts (and disrupts) calls to the Patroni API.

This commit changes the default behaviour, but I cannot think of any
scenario where the former behaviour could be useful; hence I did not
implement a user parameter to override it. Should we ever need it, we
can introduce it at a later point in time.

References: TPA-790
Signed-off-by: Gianni Ciolli <gianni.ciolli@enterprisedb.com>