Regenerate patroni dynamic config correctly
When running a deploy against an existing patroni cluster, change
patroni's dynamic configuration as required before doing postgres
actions that rely on patroni configuration being correct.

The comparison and, if necessary, changing of patroni's config now
checks the "slots" and "ignore_slots" keys for changes; these keys
are written to patroni's local yaml config at startup but because
they actually live in the dcs ("dynamic" configuration), we need to
treat them as such in order to change them.

When we ask the patroni api to restart postgres after making changes,
we now do so regardless of whether we've set a desired postgres
restart state; patroni knows what it needs to do and we can tell it
to restart postgres only if it is required.

Fixes: TPA-792, TPA-781