Add efm-pre-config hook
The efm-pre-config hook runs after efm has been installed and its
configuration directory and user have been created, but before its
configuration is generated. It can be used to install efm helper
scripts.

It runs from inside the efm/config role because the user is created as
part of the role and it is useful for the user to already exist when
installing scripts from the hook.

Fixes: TPA-791