Retry a postgres connection when it failed
We can have false-positives on slow systems if we don't retry after a
connection problem. In psycopg2 if during the connection attempt this
fails, it raises an OperationalError and the returned value is None.

We can safely keep asking if postgres is ready after taking a short
break, although, we cannot ask forever, we added a upper bondary in
which we can assume if still shows the same error, that there's really a
problem with the network and TPA won't be able to carry on.

Co-authored-by: Jonathan Renon <98327055+JonathanRenon-EDB@users.noreply.github.com>