Grant barman_role to barman user on all nodes
Since the barman user can ssh to postgres on any Postgres server (and it is not restricted to just those with the 'backup' role), it would be inconsistent to only grant the required privileges to the barman_role on backup nodes. The barman user should be the same across all nodes in the cluster.