TPA-821: New option local_etcd_only
By default, `tpaexec configure` generates `etcd` roles co-located with the `bdr` roles.

Also, by default, HARP Manager includes all `etcd` roles in the `dcs -> endpoints`, for example from `/etc/harp/config.yml`:

```
dcs:
...
  endpoints:
  - rt41391n1:2379
  - rt41391n1:2379
  - rt41391n1:2379
```

This can lead to performance issues in the HARP Manager loop, and even leading to leader loss if HARP Manager isn't able to refresh the lease lock.

To alleviate that, and also following the recommended HARP Manager configuration for when `harp_consensus_protocol: bdr` (which is to communicate only with the local BDR node using Unix sockets), this commit introduces an option called `local_etcd_only` which by default is set to `false` but, when the user enables it:

```
cluster_vars:
  ...
  harp_consensus_protocol: etcd
  local_etcd_only: true
```

Then it adds only the local `etcd` node to `dcs -> endpoints`, for example for HARP Manager:

```
dcs:
  ...
  endpoints:
  - rt41391n1:2379
```

But it doesn't change the HARP Proxy configuration as usually HARP Proxy is on a separate host, so HARP Proxy needs to communicate with all `etcd` nodes, for example:

```
dcs:
  ...
  endpoints:
  - rt41391n1:2379
  - rt41391n2:2379
  - rt41391n3:2379
```

Signed-off-by: William Ivanski <william.ivanski@enterprisedb.com>