TPA-838: Add Lightweight architecture support
Add lightweight architecture support in configure command.
based on simplified PGD-Always-ON code.
Initial support.

add a new architecture hierarchy folder

add lightweight.py to manage configure command for the new architecture.
based on pgd-always-on.py and simplified to match the requirements

ensure that lightweight is recognized as an option for --architecture
during configure.