Allow multiple IP addresses in hostnames file
If a line in the file supplied to `tpaexec configure --hostnames-from`
includes two ip addresses after the hostname, the first will be used
as the `public_ip` in config.yml and the second as the `private_ip`.
If exactly one address is included, it is still used as `ip_address`.