Expose configurable efm_user_password_encryption variable
Configures the auth-method for the efm Postgres user in pg_hba.conf
and the algorithm used for generating it's encrypted password.

In clusters deployed with compliance configured to stig, the efm
Postgres user's auth-method in pg_hba.conf will be set to
scram-sha-256 since FIPS-enabled operating systems do not allow md5
to be used.

Add documentation for this  configuration parameter.