Use client SSL certificates if provided
If the fact ssl_client_cert_dir is set to a path which contains an
ssl certificate and key for the user, each user being set up on the
client-side will have the certificate and key set up along with the
.pgpass entry. In a STIG-compliant cluster, it is an error for this not
to be the case.

Because the haproxy check-queue user has a custom PGPASSFILE set, it
also now has has custom PGSSLCERT and PGSSLKEY.

If client ssl certificates are supplied, assemble a suitable
pg_ident.conf so that they can be used in situations where the same OS
user has to connect as several db users.