Update EFM .nodes file when a new node is added
When a new EFM node is added to config.yml, it is not listed in the
allowed node host list on the existing EFM nodes in the cluster.

The task which executes  and propagates the changes
from /raw/<clustername>.properties and /raw/<clustername>.nodes is
now run when EITHER of these files are changed.

This results in the new EFM node being written to the
<clustername>.nodes file and efm cluster-status including it in the
allowed node host list.