Fix additional repository setup commands for extensions
Deploys were failing for Ubuntu, Debian and SUSE when postgis is
added as an extension because these distributions have no additional
repository setup commands to run. As a result, an empty list was
iterated and nothing was passed to the 'command' module, which
failed with no 'command given'.

These distributions were removed from the repository_setup_commands
dictionary and the task only runs when a distribution is defined
and has a list of commands to run.

Added version 3.5 to supported_postgis_versions.