Fix Postgres database creation
In order to remove modules from the list of extensions for named
databases under postgres_databases, the entire hash was modified.
This introduced a bug, since a new hash was created that ONLY
contained the database name and modified list of extensions, ignoring
all other configuration settings for the database.

This resulted in databases being created with the default parameters
rather than as configured.

To fix this, the modules are removed from the list of extensions and
the resulting list is passed to the task which runs CREATE EXTENSION.