Handle different auto.resume.* efm properties for v5 vs lower
Starting in EFM 5.0, there is no longer an auto.resume.period property.
Instead, there are two new properties, one that covers the startup of
an agent and one that covers the case of a database failure. This
change checks 'efm_version' before writing out the property(ies).

See TPA-892 for testing output. The efm change is from EFM-1614.