Upgrade to Ansible-core 2.16
Also, the requirements-ansible-[89].in file contained just
"ansible-core==<VERSION>" which was merged onto the main
requirements.in. This also implies that the only supported
version is the one included in requirements.txt since this
commit removes the mechanism to support multiple version of
ansible-core.

Updated dependencies to the latest available versions.

References: TPA-870