Fix undefined FOUNDATION_VER fetching foundation
`actions/checkout` was including the ref when cloning, however, the variable
mentioned above, sets the branch to get and it was never defined. By default,
`checkout` brings "develop" which works in most of cases but can lead to
undesired results.

Related logs from a random workflow:
  From https://github.com/EnterpriseDB/foundation-packaging
   * [new branch]      develop    -> origin/develop
Determining the checkout info
/usr/bin/git sparse-checkout disable
/usr/bin/git config --local --unset-all extensions.worktreeConfig
Checking out the ref
  /usr/bin/git checkout --progress --force -B develop refs/remotes/origin/develop
  Switched to a new branch 'develop'