Add pgbapi_dependency_packages
The changes made to add package-version option for pg-backup-api
split the pg-backup-api and dependency packages into two lists.
However, the dependency packages were not added to the
list-packages task, so were not installed.

The pkg/add_to_list interface is used to add these packages, which
are now installed alongside pg-backup-api.