Stop using a forked upload-artifact action
Revert to main upload-artifact to switch to v4 for integration tests.
We used to have a forked version that reduced the output of the action
by grouping the output.

changed the name of the artifacts to be unique, each cluster of the
integration test will now generate its own artifact instead of a single
aggregated tar ball.

We will see if this change is generating too much output in integration
tests in the future or not.