Fix pgd_{proxy,cli}_dsn_attributes defaults
Because pgd-proxy and pgd-cli are written in Go and use a Go driver,
they do not support the full set of parameter keywords supported by
libpq.

In the case a cluster has installed pgd-proxy and/or pgd-cli and has
configured bdr_client_dsn_attributes with parameters that the Go
driver does not support, two new configuration variables must be
included: pgd_proxy_dsn_attributes and pgd_cli_dsn_attributes,
containing only additional DSN parameters that the Go driver
supports.

Conversely, if pgd-proxy and pgd-cli are installed and
bdr_client_dsn_attributes does not include any Go-incompatible
parameters, the connection strings for these tools will be configured
with the attributes in bdr_client_dsn_attributes.

This amends unexpected behavior where the pgd_proxy_dsn_attributes
and pgd_cli_dsn_attributes were defaulting to empty strings when
not defined, even though the user was expecting the parameters in
bdr_client_dsn_attributes to be used.