Add support for nodes running SLES15SP6
Add docker image and AMI for SLES15SP6.

On SLES15SP6, install the systemd-sysvcompat package so that the rc-local
service will work and ensure that the corresponding script has the
correct path for each OS. Install apache2-utils so that we have
apache2ctl available. Use the correct PackageHub repository version.