Ensure URI for EDB repository is accessible
The EDB repos are set up using the Cloudsmith setup script following
the Cloudsmith documentation: piping the cURL output to bash for
execution.

However, if a user passes a nonexistent EDB_SUBSCRIPTION_TOKEN or
repository to cURL, the exit code gets silently swallowed and
replaced with a 0 because bash executes an empty input.

A request is dispatched to a repository's GPG key endpoint to ensure
a 404 response is not returned before continuing to download the
setup script, and an error message is returned to the user alerting
them that their repo token does not have access or the repository
does not exist.

Tasks related to EDB repository set up are now skipped if the
repository has already been set up.