M1 upgrade integration tests
The m1-upgrade-test action accepts an os image, a postgres flavour, a
postgres major version, a postgres package version to upgrade from, and a
failover manager. It deploys a cluster with the specified postgres package
version, then edits config.yml to remove it and upgrades the cluster. Two
workflows call this action: one workflow runs the action once, with the
supplied parameters, and the other runs it many times from a matrix.

There are four matrixes of tests that can be run:
- "basic" runs tests for each failover manager, with the other
  parameters set by the user.
- "all_os_images" runs tests for each failover manager and each os
  image, with the postgres flavour and version set by the user.
- "all_flavours" runs tests for each failover manager each postgres
  flavour, with the os image and postgres versions set by the user.
- "all_pg_versions" runs tests for each failover manager and each
  supported pg version, with the os image and postgres flavour set by
  the user and the postgres package version supplied as appropriate by
  the image. The unsupported combination of epas/repmgr is omitted.

When running from these matrixes, the downloader is used unless the
configuration is one known to not support it (SLES clusters, or RedHat
clusters using patroni as the failover manager).

Fixes: TPA-913