add support for minor postgres upgrades on patroni clusters (#487)
To upgrade postgres on a patroni cluster, we pause patroni's management
of the cluster, then upgrade and restart the replicas, then use the patroni api
to switchover to a replica, then upgrade and restart the original primary, then
switch back to it. These steps are in cluster_upgrade.yml in the patroni role,
which is included from the M1 architecture's upgrade command if the cluster
is a patroni one.

The patroni/api role now supports the switchover endpoint.

Fixes: TPA-688


---------

Co-authored-by: Jonathan Renon <jonathan.renon@enterprisedb.com>