Catch exception for --network during configure
By default, the Python standard library ipaddress package enforces
strict interpretation of the CIDR, whereby the IP used should be
the network address of the range.

Previously, any IP passed to the --network flag that contained host
bits would dump a stacktrace due to the raised ValueError.
That exception is now caught and an ArchitectureError is raised to
display a clear message to the user about the --network parameter.